Fix bugs

+++ src/main/java/com/example/content_calender/repository/ContentRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.stereotype.Repository;
 
 
-public interface ContentRepository extends ListCrudRepository<Content, Integer> {
+public interface ContentRepository extends ListCrudRepository<Content, Long> {
     // This interface extends ListCrudRepository to provide basic CRUD operations
     // for the Content entity. No additional methods are needed here as the
     // ListCrudRepository provides all necessary methods.
